Decomposed sender into smaller components
